requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.5
pycodestyle==2.9.1
pycryptodome-test-vectors==1.0.10
pycryptodomex==3.15.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.6
